also moved tuple class into the model package
