+++ app/src/main/java/com/goayo/debtify/Database/DbDebtsFetcher.java
@@ -1,7 +1,7 @@
 package com.goayo.debtify.Database;
 
 import com.goayo.debtify.model.IDebtSplitStrategy;
-import com.goayo.debtify.Tuple;
+import com.goayo.debtify.model.Tuple;
 import com.goayo.debtify.model.IUserData;
 import com.google.gson.Gson;
