+++ app/src/main/java/com/goayo/debtify/model/Ledger.java
@@ -1,8 +1,6 @@
 package com.goayo.debtify.model;
 
 
-import com.goayo.debtify.Tuple;
-
 import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
