+++ app/src/main/java/com/goayo/debtify/model/NoSplitStrategy.java
@@ -1,7 +1,5 @@
 package com.goayo.debtify.model;
 
-import com.goayo.debtify.Tuple;
-
 import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
